fix: move TABLES_WITH_ASSETS to constants and fix BalanceType imports
